# v1/wwwroot/openapi.yaml
openapi: 3.0.3
info:
  title: ASP.NET Core | Entity Framework Core
  description: |-
    Данное Rest-API выполняет роль backend части для [NetheriteTrade](https://t.me/netheritetrade_bot/netheriteapp) и может быть использовано для
    1) Добавление задач
    2) Регистрации пользователей
    3) Майнинга виртуальной валюты
  version: 1.0.11
servers:
  - url: https://5.35.81.57
  - url: http://localhost:5068
...
paths:
...
    /register:
      post:
        tags:
          - User
        summary: Регистрация нового пользователя
        description: Создание нового пользователя в системе
        operationId: RegisterUser
        parameters:
          - name: telegram_id
            in: query
            description: Telegram ID пользователя (не путать с @username)
            required: true
            schema:
              type: string
              example: "123456789"
          - name: telegram_name
            in: query
            description: Имя пользователя в Telegram
            required: true
            schema:
              type: string
              example: "Иван Иванов"
          - name: premium
            in: query
            description: Простой аккаунт или Премиум аккаунт
            required: true
            schema:
              type: boolean
              example: false
          - name: geo
            in: query
            description: Географическое положение пользователя
            required: true
            schema:
              type: string
              example: "Москва"
          - name: invitedId
            in: query
            description: ID приглашающего пользователя
            required: false
            schema:
              type: string
              nullable: true
              example: null
          - name: walletId
            in: query
            description: ID кошелька пользователя
            required: true
            schema:
              type: string
              example: "WALLET_12345"
          - name: tonUserId
            in: query
            description: ID пользователя в Ton
            required: true
            schema:
              type: string
              example: "TON_USER_67890"
        responses:
          '200':
            description: Регистрация прошла успешно
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: integer
                      example: 200
                    callback:
                      type: object
                      properties:
                        msg:
                          type: string
                          example: "Registration successful"
                        id:
                          type: string
                          format: uuid
                          example: "f397aa41-cfa3-481a-b5c8-dec8de1f3ba6"
          '400':
            description: Неправильный тип ID
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: integer
                      example: 400
                    callback:
                      type: string
                      example: "Invalid telegram_id format"
          '409':
            description: Пользователь с таким ID уже существует
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: integer
                      example: 409
                    callback:
                      type: string
                      example: "This telegram_id already exists"
    /user/{telegram_id}:
      get:
        tags:
          - User
        summary: Получить информацию о пользователе по Telegram ID
        description: Возвращает данные о пользователе, идентифицированном по его Telegram ID
        operationId: GetUserByTelegramId
        parameters:
          - name: telegram_id
            in: path
            description: Telegram ID пользователя (не путать с @username)
            required: true
            schema:
              type: string
              example: "123456789"
        responses:
          '200':
            description: Информация о пользователе успешно получена
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: integer
                      example: 200
                    callback:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          example: "f397aa41-cfa3-481a-b5c8-dec8de1f3ba6"
                        balance:
                          type: number
                          example: 1025
                        tonUserId:
                          type: string
                          example: "TON_USER_67890"
                        walletId:
                          type: string
                          example: "WALLET_12345"
          '404':
            description: Пользователь с указанным Telegram ID не найден
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    code:
                      type: integer
                      example: 404
                    callback:
                      type: string
                      example: "User not found"
...
    /user/{telegram_id}/complete/{task_id}:
      post:
        tags:
          - User
        summary: Завершить задачу и обновить баланс пользователя
        description: Завершает задачу Task и обновляет баланс пользователя на основе назначенного администратором вознаграждения
        operationId: CompleteTask
        parameters:
          - name: telegram_id
            in: path
            description: Telegram ID пользователя (не путать с @username)
            required: true
            schema:
              type: string
              example: "123456789"
          - name: task_id
            in: path
            description: UUID существующей задачи
            required: true
            schema:
              type: string
              example: "f397aa41-cfa3-481a-b5c8-dec8de1f3ba6"
        responses:
          '200':
            description: Задача успешно завершена
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: integer
                      example: 200
                    message:
                      type: string
                      example: "Task completed successfully"
          '400':
            description: Неверный формат ID задачи
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: integer
                      example: 400
                    message:
                      type: string
                      example: "Invalid task ID format"
          '404':
            description: Задача или пользователь не найдены
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: integer
                      example: 404
                    message:
                      type: string
                      example: "Task not found"
          '500':
            description: Неожиданная ошибка
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: integer
                      example: 500
                    message:
                      type: string
                      example: "An unexpected error occurred"
...